pyproject.toml
[tool.poetry]
name = "flow-pinecone-chatgpt"
version = "0.1.0"
description = "Example scripts for querying ChatGPT using retrieval augmentation with a Pinecone database populated by a Flow materialization"
authors = ["Will Baker <wbaker@gmail.com>"]
readme = "README.md"
packages = [{include = "flow_pinecone_chatgpt"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.7"
pinecone-client = "^2.2.1"
langchain = "^0.0.186"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
